page-5-maintenance-services/backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database & ORM
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.1
